Added DIR helper to netstorage.baseservice.Binding
`Binding.dir()` returns a list of dicts with these keys:
'type', 'name', 'mtime', 'path' and if applicable, these keys:
'size' (only for file type entries), 'target' (only for 'symlink' type entries)

'path' is computed from the `path` parameter passed to `dir` and the 'name' key.
If the `path` parameter is left empty, the 'path' key is the same as the 'name' key.
The 'size' and 'mtime' keys are converted to proper Python values (int and datetime,
respectively). If an error occurs during value conversion, they are left as strings.